ゲージ用シェーダーの追加 ディファード用画像シェーダーの追加 被写界深度の削除

+++ MyGAME/GameTemplate/Game/Path.h
@@ -40,7 +40,7 @@ public:
 	 */
 	int GetNumberPoint()
 	{
-		return m_pointList.size();
+		return static_cast<int>(m_pointList.size());
 	}
 	/**
 	 * \brief Å‰‚Ìƒ|ƒCƒ“ƒg‚ğæ“¾.
